+++ src/core/domain/model/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod cluster_resource;
 pub(crate) mod proxmox_auth;
 pub(crate) mod proxmox_connection;